Moved withStartingIndicator calls to pages index
