+++ web/ui/react-app/src/pages/index.ts
@@ -7,5 +7,16 @@ import Status from './status/Status';
 import Targets from './targets/Targets';
 import PanelList from './graph/PanelList';
 import TSDBStatus from './tsdbStatus/TSDBStatus';
+import { withStartingIndicator } from '../components/withStartingIndicator';
 
-export { Alerts, Config, Flags, Rules, ServiceDiscovery, Status, Targets, TSDBStatus, PanelList };
+const AlertsPage = withStartingIndicator(Alerts);
+const ConfigPage = withStartingIndicator(Config);
+const FlagsPage = withStartingIndicator(Flags);
+const RulesPage = withStartingIndicator(Rules);
+const ServiceDiscoveryPage = withStartingIndicator(ServiceDiscovery);
+const StatusPage = withStartingIndicator(Status);
+const TSDBStatusPage = withStartingIndicator(TSDBStatus);
+const TargetsPage = withStartingIndicator(Targets);
+const PanelListPage = withStartingIndicator(PanelList);
+
+export { AlertsPage, ConfigPage, FlagsPage, RulesPage, ServiceDiscoveryPage, StatusPage, TSDBStatusPage, TargetsPage, PanelListPage };
